refactor: move stuff to darkgrpc
