+++ darkapis/darkgrpc/guns.go
@@ -4,7 +4,6 @@ import (
 	"context"
 
 	pb "github.com/darklab8/fl-darkstat/darkapis/darkgrpc/statproto"
-	"github.com/darklab8/fl-darkstat/darkapis/services"
 	"github.com/darklab8/fl-darkstat/darkstat/configs_export"
 )
 
@@ -19,7 +18,7 @@ func (s *Server) GetGuns(_ context.Context, in *pb.GetGunsInput) (*pb.GetGunsRep
 	} else {
 		input = s.app_data.Configs.Guns
 	}
-	input = services.FilterNicknames(in.FilterNicknames, input)
+	input = FilterNicknames(in.FilterNicknames, input)
 
 	var items []*pb.Gun
 	for _, item := range input {
@@ -40,7 +39,7 @@ func (s *Server) GetMissiles(_ context.Context, in *pb.GetGunsInput) (*pb.GetGun
 	} else {
 		input = s.app_data.Configs.Missiles
 	}
-	input = services.FilterNicknames(in.FilterNicknames, input)
+	input = FilterNicknames(in.FilterNicknames, input)
 
 	var items []*pb.Gun
 	for _, item := range input {
