+++ darkapis/darkgrpc/misc_hashes.go
@@ -1,6 +1,8 @@
-package services
+package darkgrpc
 
 import (
+	"context"
+
 	"runtime"
 	"strings"
 	"sync"
@@ -8,6 +10,7 @@ import (
 	"github.com/darklab8/fl-darkstat/configs/configs_mapped/freelancer_mapped/data_mapped/initialworld/flhash"
 	"github.com/darklab8/fl-darkstat/configs/configs_mapped/parserutils/filefind"
 	"github.com/darklab8/fl-darkstat/configs/configs_mapped/parserutils/iniload"
+	pb "github.com/darklab8/fl-darkstat/darkapis/darkgrpc/statproto"
 	"github.com/darklab8/fl-darkstat/darkstat/appdata"
 	"github.com/darklab8/fl-darkstat/darkstat/settings"
 )
@@ -75,4 +78,25 @@ func GetHashesData(app_data *appdata.AppData) map[string]Hash {
 		}
 	}
 	return hashes
+}
+
+func (s *Server) GetHashes(_ context.Context, in *pb.Empty) (*pb.GetHashesReply, error) {
+	if s.app_data != nil {
+		s.app_data.Lock()
+		defer s.app_data.Unlock()
+	}
+
+	answer := &pb.GetHashesReply{HashesByNick: make(map[string]*pb.Hash)}
+
+	hashes := GetHashesData(s.app_data)
+
+	for key, hash := range hashes {
+		answer.HashesByNick[key] = &pb.Hash{
+			Int32:  hash.Int32,
+			Uint32: hash.Uint32,
+			Hex:    hash.Hex,
+		}
+	}
+
+	return answer, nil
 }
