+++ darkapis/darkgrpc/shared.go
@@ -3,7 +3,6 @@ package darkgrpc
 import (
 	"github.com/darklab8/fl-darkstat/configs/cfg"
 	pb "github.com/darklab8/fl-darkstat/darkapis/darkgrpc/statproto"
-	"github.com/darklab8/fl-darkstat/darkapis/services"
 	"github.com/darklab8/fl-darkstat/darkstat/configs_export"
 )
 
@@ -64,7 +63,7 @@ func NewPos(pos *cfg.Vector) *pb.Pos {
 
 func NewBases(Bases map[cfg.BaseUniNick]*configs_export.MarketGood, filter_market_good_category []string) map[string]*pb.MarketGood {
 	result := make(map[string]*pb.MarketGood)
-	for key, item := range services.FilterMarketGoodCategory(filter_market_good_category, Bases) {
+	for key, item := range FilterMarketGoodCategory(filter_market_good_category, Bases) {
 		result[string(key)] = NewMarketGood(item)
 	}
 	return result
