+++ darkapis/darkhttp/bases.go
@@ -3,9 +3,9 @@ package darkhttp
 import (
 	"net/http"
 
+	"github.com/darklab8/fl-darkstat/darkapis/darkgrpc"
 	pb "github.com/darklab8/fl-darkstat/darkapis/darkgrpc/statproto"
 	"github.com/darklab8/fl-darkstat/darkapis/darkhttp/apiutils"
-	"github.com/darklab8/fl-darkstat/darkapis/services"
 	"github.com/darklab8/fl-darkstat/darkcore/web"
 	"github.com/darklab8/fl-darkstat/darkcore/web/registry"
 	"github.com/darklab8/fl-darkstat/darkstat/configs_export"
@@ -53,7 +53,7 @@ func GetBases(webapp *web.Web, api *Api) *registry.Endpoint {
 			} else {
 				result = api.app_data.Configs.Bases
 			}
-			result = services.FilterNicknames(in.FilterNicknames, result)
+			result = darkgrpc.FilterNicknames(in.FilterNicknames, result)
 
 			var output []*Base
 			for _, item := range result {
@@ -61,7 +61,7 @@ func GetBases(webapp *web.Web, api *Api) *registry.Endpoint {
 					Base: item,
 				}
 				if include_market_goods {
-					for _, good := range services.FilterMarketGoodCategory(in.FilterMarketGoodCategory, item.MarketGoodsPerNick) {
+					for _, good := range darkgrpc.FilterMarketGoodCategory(in.FilterMarketGoodCategory, item.MarketGoodsPerNick) {
 						answer.MarketGoods = append(answer.MarketGoods, good)
 					}
 				}
@@ -109,7 +109,7 @@ func GetOreFields(webapp *web.Web, api *Api) *registry.Endpoint {
 			} else {
 				result = api.app_data.Configs.MiningOperations
 			}
-			result = services.FilterNicknames(in.FilterNicknames, result)
+			result = darkgrpc.FilterNicknames(in.FilterNicknames, result)
 
 			var output []*Base
 			for _, item := range result {
@@ -117,7 +117,7 @@ func GetOreFields(webapp *web.Web, api *Api) *registry.Endpoint {
 					Base: item,
 				}
 				if include_market_goods {
-					for _, good := range services.FilterMarketGoodCategory(in.FilterMarketGoodCategory, item.MarketGoodsPerNick) {
+					for _, good := range darkgrpc.FilterMarketGoodCategory(in.FilterMarketGoodCategory, item.MarketGoodsPerNick) {
 						answer.MarketGoods = append(answer.MarketGoods, good)
 					}
 				}
@@ -154,7 +154,7 @@ func GetPoBBases(webapp *web.Web, api *Api) *registry.Endpoint {
 				include_market_goods = true
 			}
 			var result []*configs_export.Base = api.app_data.Configs.PoBsToBases(api.app_data.Configs.PoBs)
-			result = services.FilterNicknames(in.FilterNicknames, result)
+			result = darkgrpc.FilterNicknames(in.FilterNicknames, result)
 
 			var output []*Base
 			for _, item := range result {
@@ -162,7 +162,7 @@ func GetPoBBases(webapp *web.Web, api *Api) *registry.Endpoint {
 					Base: item,
 				}
 				if include_market_goods {
-					for _, good := range services.FilterMarketGoodCategory(in.FilterMarketGoodCategory, item.MarketGoodsPerNick) {
+					for _, good := range darkgrpc.FilterMarketGoodCategory(in.FilterMarketGoodCategory, item.MarketGoodsPerNick) {
 						answer.MarketGoods = append(answer.MarketGoods, good)
 					}
 				}
